+++ example/lib/src/view.dart
@@ -1,5 +1,5 @@
 // The fluttery framework.
-export 'package:fluttery_framework/view.dart' hide AppMenu;
+export 'package:fluttery_framework/view.dart' hide AppDrawer, AppMenu;
 
 // The UI at the app level.
 export 'package:fluttery_framework_example/src/app/view.dart' hide ISOSpinner;